# backend/api_spec.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "ORBIT API",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/cycles/": {
            "get": {
                "tags": [
                    "cycles"
                ],
                "summary": "List Cycles",
                "operationId": "list_cycles_api_cycles__get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/TestCycle"
                                    },
                                    "type": "array",
                                    "title": "Response List Cycles Api Cycles  Get"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "cycles"
                ],
                "summary": "Create Cycle",
                "operationId": "create_cycle_api_cycles__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TestCycleCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TestCycle"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/cycles/{cycle_id}": {
            "get": {
                "tags": [
                    "cycles"
                ],
                "summary": "Get Cycle",
                "operationId": "get_cycle_api_cycles__cycle_id__get",
                "parameters": [
                    {
                        "name": "cycle_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Cycle Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TestCycle"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "cycles"
                ],
                "summary": "Update Cycle",
                "operationId": "update_cycle_api_cycles__cycle_id__put",
                "parameters": [
                    {
                        "name": "cycle_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Cycle Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TestCycleUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TestCycle"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "cycles"
                ],
                "summary": "Delete Cycle",
                "operationId": "delete_cycle_api_cycles__cycle_id__delete",
                "parameters": [
                    {
                        "name": "cycle_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Cycle Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/cycles/{cycle_id}/executions": {
            "get": {
                "tags": [
                    "cycles"
                ],
                "summary": "List Cycle Executions",
                "operationId": "list_cycle_executions_api_cycles__cycle_id__executions_get",
                "parameters": [
                    {
                        "name": "cycle_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Cycle Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TestExecution"
                                    },
                                    "title": "Response List Cycle Executions Api Cycles  Cycle Id  Executions Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "cycles"
                ],
                "summary": "Add Execution To Cycle",
                "operationId": "add_execution_to_cycle_api_cycles__cycle_id__executions_post",
                "parameters": [
                    {
                        "name": "cycle_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Cycle Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CycleExecutionLink"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "cycles"
                ],
                "summary": "Remove Execution From Cycle",
                "operationId": "remove_execution_from_cycle_api_cycles__cycle_id__executions_delete",
                "parameters": [
                    {
                        "name": "cycle_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Cycle Id"
                        }
                    },
                    {
                        "name": "execution_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Execution Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CycleExecutionLink": {
                "properties": {
                    "execution_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Execution Id"
                    }
                },
                "type": "object",
                "required": [
                    "execution_id"
                ],
                "title": "CycleExecutionLink"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "TestCycle": {
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "title": "Name"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 1024
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description"
                    },
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "id",
                    "created_at"
                ],
                "title": "TestCycle"
            },
            "TestCycleCreate": {
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "title": "Name"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 1024
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description"
                    }
                },
                "type": "object",
                "required": [
                    "name"
                ],
                "title": "TestCycleCreate"
            },
            "TestCycleUpdate": {
                "properties": {
                    "name": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 255
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Name"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 1024
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description"
                    }
                },
                "type": "object",
                "title": "TestCycleUpdate"
            },
            "TestExecution": {
                "properties": {
                    "status": {
                        "$ref": "#/components/schemas/TestExecutionStatus"
                    },
                    "started_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Started At"
                    },
                    "completed_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Completed At"
                    },
                    "duration_ms": {
                        "anyOf": [
                            {
                                "type": "integer",
                                "minimum": 0.0
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Duration Ms"
                    },
                    "result_data": {
                        "anyOf": [
                            {
                                "additionalProperties": true,
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Result Data"
                    },
                    "logs_url": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 1024
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Logs Url"
                    },
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "test_case_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Test Case Id"
                    }
                },
                "type": "object",
                "required": [
                    "status",
                    "started_at",
                    "id",
                    "test_case_id"
                ],
                "title": "TestExecution"
            },
...
            "TestExecutionStatus": {
                "type": "string",
                "enum": [
                    "passed",
                    "failed",
                    "blocked"
                ],
                "title": "TestExecutionStatus"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}